Initial Commit
This project receives input from a file that contains work orders(name - number(amount of work orders)). It then takes that input, adds up the work orders for entries of the same name, sorts from largest amount of work orders to least using qsort, and then outputs that information to a separate file.